Create an abstraction above arrays.
CUDA requires manipulation with blocks located
in the device memory. For debugging purposes and
so on, it is convenient to have an arrayish
interface for them, even if it is a lot slower
than the ordinary arrays. This also provides a
general framework for the actually useful bulk
copy methods.