Relax GPU kernel argument type restrictions a bit.
Allow using any buffer as an argument to a CUDA
kernel by means of allocating an acceptable
substitute and copying data. This might make
debugging a bit more convenient.

Side effect analysis results are used to decide
which of the buffers need to be copied back from
the GPU memory.