Implement a way to call gpu-function objects from REPL.
Using global GPU functions normally requires creating
a module. This is not convenient for REPL, so implement
a helper function that does that automatically.

Since GPU kernels require all formal argument types
to be declared, the actual arguments are inspected to
fill in the missing information. This makes test
calls quite expensive.