Fix the finalization regression test on SBCL.
SBCL conservatively pins down the whole page for every pointer
found on the stack or in the registers, so allocate a few pages
of garbage before and after the test object. Also use recursion
and scrub-control-stack to be on the safe side.

The foo variable is used so that SBCL's dead code analysis
won't remove the array creation calls.